Move xml_open to xfile
