+++ tableaudocumentapi/xfile.py
@@ -3,8 +3,31 @@ import os
 import shutil
 import tempfile
 import zipfile
-
 import xml.etree.ElementTree as ET
+
+try:
+    from distutils2.version import NormalizedVersion as Version
+except ImportError:
+    from distutils.version import LooseVersion as Version
+
+MIN_SUPPORTED_VERSION = Version("9.0")
+
+
+class TableauVersionNotSupportedException(Exception):
+    pass
+
+
+def xml_open(filename):
+    # Determine if this is a twb or twbx and get the xml root
+    if zipfile.is_zipfile(filename):
+        tree = get_xml_from_archive(filename)
+    else:
+        tree = ET.parse(filename)
+    file_version = Version(tree.getroot().attrib.get('version', '0.0'))
+    print(file_version, MIN_SUPPORTED_VERSION)
+    if file_version < MIN_SUPPORTED_VERSION:
+        raise TableauVersionNotSupportedException(file_version)
+    return tree
 
 
 @contextlib.contextmanager
@@ -16,10 +39,10 @@ def temporary_directory(*args, **kwargs):
         shutil.rmtree(d)
 
 
-def find_file_in_zip(zip):
-    for filename in zip.namelist():
+def find_file_in_zip(zip_file):
+    for filename in zip_file.namelist():
         try:
-            with zip.open(filename) as xml_candidate:
+            with zip_file.open(filename) as xml_candidate:
                 ET.parse(xml_candidate).getroot().tag in (
                     'workbook', 'datasource')
                 return filename
@@ -36,14 +59,14 @@ def get_xml_from_archive(filename):
     return xml_tree
 
 
-def build_archive_file(archive_contents, zip):
+def build_archive_file(archive_contents, zip_file):
     for root_dir, _, files in os.walk(archive_contents):
         relative_dir = os.path.relpath(root_dir, archive_contents)
         for f in files:
             temp_file_full_path = os.path.join(
                 archive_contents, relative_dir, f)
             zipname = os.path.join(relative_dir, f)
-            zip.write(temp_file_full_path, arcname=zipname)
+            zip_file.write(temp_file_full_path, arcname=zipname)
 
 
 def save_into_archive(xml_tree, filename, new_filename=None):
